allow access to context in type check
